+++ src/services/api/save-game-result.ts
@@ -1,5 +1,5 @@
 import { getState } from '../../app/state/store.ts';
-import { supabase } from '../supabaseClient.ts';
+import { supabase } from '../supabase-client.ts';
 import { withApiErrorHandling } from '../../shared/helpers/request-error.ts';
 
 export function saveGameResult() {
