+++ src/services/auth-service.ts
@@ -1,4 +1,4 @@
-import { supabase } from './supabaseClient';
+import { supabase } from './supabase-client';
 import type {
   User,
   AuthSession,
